+++ src/math/lp/nex.h
