+++ src/math/lp/nla_core.cpp
@@ -19,7 +19,7 @@ Revision History:
 --*/
 #include "math/lp/nla_core.h"
 #include "math/lp/factorization_factory_imp.h"
-#include "math/lp/nla_expr.h"
+#include "math/lp/nex.h"
 namespace nla {
 
 core::core(lp::lar_solver& s, reslimit & lim) :
